backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sse-starlette==1.8.2
anthropic==0.79.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
watchfiles==0.21.0
pytest==7.4.4
pytest-asyncio==0.23.3
